Output for each band
Now, you can put in a cube or a single fits file and it generates maps per frequency band. If you put in 1 image as input. The code uses that image for all bands. But you can put in an image per band now. So you can have spectral effects.