Units
I have added a unit tag. Currently, we can work in Kelvin Rayleigh Jeans and or Jy/pixel. Under the hood, everything is KRJ so if units is specified as Jy/pixel, it converts everything to KRJ, but the outputs are all given in the specified units. So if Jy/pixel is specified, the saved fits files will output something in Jy/pixel.
The units of the cmb should also be correct now.