map_freq is wrong
seems to overwrite it to 150 somehwere, can't find it. We should change that the defeautlt value for the map parameters are that of the telescope.

Also, if not specifing the detector_config in the sim class it just adds the defeault value instead of the one that is specified in array.


Former-commit-id: 588d33f0af863aa8d7bceef72a769bfb6c25e1f6 [formerly f8191f949cd743c5bc69eab2880c3fb64b69140e] [formerly 5dbfe2c27d4fcbb51ac6fe43d85a7dd4c3af5d05 [formerly 9eedc207bae9ad08c5fd6a15defe229751587657]]
Former-commit-id: 4b216649ed4d19b24286014a949f8c5f5ee38f30 [formerly ef7afd5224ce6abdc2b655da03ad2d00d0330421]
Former-commit-id: de9c870af075afcfeb51b8c188be0744d1f13ca0