map_freq is wrong
seems to overwrite it to 150 somehwere, can't find it. We should change that the defeautlt value for the map parameters are that of the telescope.

Also, if not specifing the detector_config in the sim class it just adds the defeault value instead of the one that is specified in array.