added units to map
default is K_RJ. If Jy/pixel is specified it converts it directly to K_RJ. when calling to_fits, it converts the data back to Jy/pixel if that was initially specified.

Also, the notebook had the key "filter" in mappers. This is not a post-processing thing. The key should be highpass.


Former-commit-id: 9111d216c03fa5ff20ef5a1bc19fa7184890f0a2 [formerly d2436bcebda257916822d6378705e0840ac0c9a8] [formerly 9111d216c03fa5ff20ef5a1bc19fa7184890f0a2 [formerly d2436bcebda257916822d6378705e0840ac0c9a8] [formerly 7064162d943fe6417e9b32065958d414911610f9]]
Former-commit-id: 369169f229793ec63a09a6a3efede7b48203ca0f
Former-commit-id: f967d138b6ecd2a17d36d89f6d0537bb12987568
Former-commit-id: 22096f4b5237061bec895cdc335928cac22b7ec1