added units to map
default is K_RJ. If Jy/pixel is specified it converts it directly to K_RJ. when calling to_fits, it converts the data back to Jy/pixel if that was initially specified.

Also, the notebook had the key "filter" in mappers. This is not a post-processing thing. The key should be highpass.