added units to map
default is K_RJ. If Jy/pixel is specified it converts it directly to K_RJ. when calling to_fits, it converts the data back to Jy/pixel if that was initially specified.

Also, the notebook had the key "filter" in mappers. This is not a post-processing thing. The key should be highpass.


Former-commit-id: d2436bcebda257916822d6378705e0840ac0c9a8 [formerly 7064162d943fe6417e9b32065958d414911610f9]
Former-commit-id: 369169f229793ec63a09a6a3efede7b48203ca0f